[topgen] Remove code duplication from TopGenC/Rust
Both of these classes implement the same functionality, either for C
or Rust by using the CEnum/CArrayMapping or RustEnum/RustArrayMapping
classes. The business logic is identical though.

This PR removes this duplication by factoring out a common base class
TopGen. This class is parametrized by using the types from before.
