+++ util/topgen/BUILD
@@ -56,6 +56,7 @@ py_library(
         "rust.py",
     ],
     deps = [
+        "//util:version_file",
         "//util/reggen:inter_signal",
         "//util/reggen:ip_block",
         "//util/reggen:validate",
